Fixed edited JSON saving
