+++ ui/src/components/workflow/WorkflowMetaDetails.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import { Tabs, Tab } from 'react-bootstrap';
 import { connect } from 'react-redux';
-import { getWorkflowMetaDetails, setWorkflowMetaDetails } from '../../actions/WorkflowActions';
+import { getWorkflowMetaDetails, updateWorkflow } from '../../actions/WorkflowActions';
 import WorkflowMetaDia from './WorkflowMetaDia';
 import WorkflowMetaInput from './WorkflowMetaInput';
 import { getInputs, getDetails } from '../../actions/JSONActions';
-import unescapeJs from 'unescape-js';
+import JSONTab from './JSONTab';
 
 class WorkflowMetaDetails extends Component {
 
@@ -72,62 +72,6 @@ class WorkflowMetaDetails extends Component {
       </div>
     );
   }
-};
-
-class JSONTab extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      editingJSON: false,
-      wfs: this.props.wfs,
-      isNotParsable: false
-    }
-    this.editJSONswitch = this.editJSONswitch.bind(this);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.setState({
-      wfs: nextProps.wfs
-    })
-  }
-
-  editJSONswitch(e, which) {
-    this.state.isNotParsable = false;
-    if(which == 1) {
-      if(this.state.editingJSON) {
-        this.setState({wfs: this.editor.innerText});
-        try {
-          JSON.parse(unescapeJs(this.editor.innerText)).tasks;
-        } catch(e) {
-          this.setState({isNotParsable : true});
-        }
-        let toBeSent = JSON.parse(unescapeJs(this.editor.innerText)).tasks;
-        console.log(setWorkflowMetaDetails(this.props.name, this.props.version, toBeSent));
-      } else {
-        this.editor.focus();
-      }
-    } else {
-      this.editor.innerText = this.state.wfs;
-    }
-    this.setState({
-      editingJSON: !this.state.editingJSON
-    });
-    this.forceUpdate();
-  }
-
-  render() {
-    return(
-        <div>
-          <button className="btn btn-primary" onClick={(e) => this.editJSONswitch(e, 1)} style={{marginTop:'5px', marginBottom: '5px'}} >{this.state.editingJSON ? 'Save' : 'Edit'}</button>
-          <button className="btn btn-default" onClick={(e) => this.editJSONswitch(e, 2)} style={{marginTop:'5px', marginBottom: '5px', marginLeft: '5px', display: this.state.editingJSON ? 'inline-block' : 'none'}} >Cancel</button>
-          <div style={{marginTop: '10px', display: this.state.isNotParsable ? "block" : "none"}} className="alert alert-warning" role="alert">{this.state.isNotParsable ? "Could not parse JSON. Is the syntax correct?" : ""}</div>
-          <pre ref={elem => this.editor = elem} className={this.state.editingJSON ? 'editingPre' : ''} contentEditable={this.state.editingJSON}>
-            {this.state.wfs}
-          </pre>
-        </div>
-    
-    )
-  };
 };
 
 export default connect(state => state.workflow)(WorkflowMetaDetails);
